+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Security - in case of vulnerabilities.
 - Bug in showing inactive news to logged users
 - Visual glitches in IE and mobile version
 - Bugs in showing system information about records
+- Dubious column names in documents and organisations
 
 ## [0.6.2] - 2018-10-17
 ### Fixed